+++ src/com/plasstech/lang/d2/type/VarType.java
@@ -12,8 +12,7 @@ public interface VarType {
       };
   VarType BOOL = new SimpleType("BOOL", 1);
   VarType BYTE = new SimpleType("BYTE", 1);
-  // technically this is a "double" but /shrug.
-  VarType FLOAT = new SimpleType("FLOAT", 8);
+  VarType DOUBLE = new SimpleType("DOUBLE", 8);
   VarType VOID = new SimpleType("VOID");
   VarType PROC = new SimpleType("PROC");
   VarType NULL = new NullType();
